requirements.txt
pandas==2.3.2
python-dotenv==1.1.1
Requests==2.32.5
SQLAlchemy==1.4.49

